+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## V2.2 - 05/01/2026
+### Añadido
+ - Carpeta **docs** para archivos de ejemlpo.
+### Modificado
+ - **instalador.sh**: Se implementa el uso de las banderas --help; --version; -x o --express; -x--help; -vm.  
+ - Se modifica el script para permitir instalación normal o instalacion express(tarda 1 minuto menos aprox) y para mostrar IP de la máquina virtual si esta siendo instalado en una, para poder acceder desde máquina host.  
+ - Se cambia el formato **txt** del documento **informe-instalacion.txt** a **html**.  
+ 
 ## V2.1 - 02/01/2026
 ### Añadido
  - Este **CHANGELOG.md**